Add feature util to make click captured feature and docs
